--- server_ADMIN/servicio.py
@@ -1,10 +1,0 @@
-from flask import Flask
-import controlador
-
-app = Flask(__name__)
-
-@app.route("/<prompt>")
-def main(prompt):
-    return controlador.main(prompt)
-
-app.run(port=5001)
